+++ crates/indexer/src/lib.rs
@@ -4,7 +4,7 @@
 // without even the implied warranty of MERCHANTABILITY
 // or FITNESS FOR A PARTICULAR PURPOSE.
 
-pub mod delayed_handlers;
+pub mod callback_queue;
 mod indexer;
 pub mod models;
 mod repo;
